requirements.txt
Flask>=3.0.0
Pillow>=9.0.0
img2pdf>=0.4.0